server/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
python-jose[cryptography] 
passlib 
python-multipart
python-dotenv
httpx
langchain
langchain-groq